gh meta: add setup.py
